pull out some trace initialization code out into trace-util.js, add fastify stub
