+++ lib/instrumentation.js
@@ -8,17 +8,18 @@ const shimmer = require("shimmer"),
   debug = require("debug")(`${pkg.name}:instrumentation`);
 
 const instrumentations = [
+  "bluebird",
+  "child_process",
   "express",
-  "react-dom/server",
-  "mysql2",
+  "fastify",
   "http",
   "https",
   "mongodb",
   "mongoose",
   "mpromise",
-  "bluebird",
+  "mysql2",
+  "react-dom/server",
   "sequelize",
-  "child_process",
 ];
 
 let enabledInstrumentations = new Set(instrumentations);
